Standardize stopword behavior to treat empty lists as non-null
When NoneType is passed to stopwords, the TextCleaner defaults to English stopwords. However, previously, if an empty list was passed, this was also being treated as NoneType, making it impossible to disable stopwords entirely. This update will treat empty lists as non-null and apply the empty list the same way non-empty lists are handled. This is more intuitive behavior - when a user passes a list of stopwords to the stopwords parameter, it is treated as a custom list of stopwords that overrides the default English stopword list. Passing an empty list should therefore be taken as specifying that no stopwords should be used.